Added a new sync phase
When a client connects to a server there will now be a new syncing phase
before the user gets control. In this phase, every detectable object
will update its position on the server and store a flag on the client.
This flag will be set to true if the object is within detect distance
and false if it isn't. On the server if this flag is false the
gameobject will get disabled and enabled if it is true. In play mode
when this flag changes the client will tell the server.